requirements.txt
asgiref==3.8.1
Django==5.1.6
sqlparse==0.5.3
tzdata==2025.1
gunicorn==23.0.0
waitress==3.0.2
django-ckeditor == 6.1.0
whitenoise == 6.11.0
Pillow==9.5.0

